requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
pymongo==4.7.2
sentence-transformers==3.0.1
numpy==1.26.4
tqdm==4.66.4
nltk==3.8.1
openai==1.40.2
requests==2.32.3
